Add queso methods to gslnumericvector/matrix
We had an overload that relied on the last arg to the matrix ctor being
a double.  It got implicitly converted to an int thus creating a matrix
of the wrong size.  That's been fixed now.

+++ src/core/src/GslNumericBlockMatrix.C
@@ -755,7 +755,7 @@ GslNumericBlockMatrix<T>::getBlock(unsigned int i) const
 
   GslSparseMatrix<T> answer(internal_answer.env(),
                             internal_answer.map(),
-                            internal_answer.map().NumGlobalElements());
+                            (unsigned int)internal_answer.map().NumGlobalElements());
 
   return answer;
 }
